Fix negative duration in Activity
Fix overflow due to cast from int to short while extending the maximum
duration to the one in PatternActivity.

+++ mobitopp/src/test/java/edu/kit/ifv/mobitopp/simulation/activityschedule/ActivityPeriodTest.java
@@ -52,8 +52,8 @@ public class ActivityPeriodTest {
 											(byte)activityNr,
 											activityType,
 											startDate,
-											(short)duration,
-											(short)tripDuration
+											duration,
+											tripDuration
 										);
 
 		return act;
